backend/requirements.txt
# ==========================================
# 1. 核心 Web 框架与服务器
# ==========================================
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# ==========================================
# 2. 数据库与 ORM 映射
# ==========================================
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# ==========================================
# 3. 配置管理与环境变量
# ==========================================
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ==========================================
# 4. 安全、认证与加密体系
# ==========================================
passlib[bcrypt]==1.7.4
# ⚠️ 锁定 bcrypt 版本，防止 passlib 内部验证时触发 72 bytes 报错
bcrypt==3.2.2
python-jose[cryptography]==3.3.0

# ==========================================
# 5. 网络请求与外部 API 交互
# ==========================================
requests==2.32.3
httpx==0.28.1

# ==========================================
# 6. AI 算法与多模态数据处理 (核心新增)
# ==========================================
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
opencv-python==4.10.0.84
Pillow==11.0.0

# ==========================================
# 7. 单元测试框架
# ==========================================
pytest==8.3.4
pytest-asyncio==0.24.0